Add folder config and  move configurations files
