+++ src/server/config/configEnv.js
@@ -2,4 +2,4 @@
 import path from "path";
 import { config } from "dotenv";
 
-config({ path: path.join(__dirname, "/../../.env") });
+config({ path: path.join(__dirname, ".env") });
